Phase 5 & 6: documentation, code organization, unsafe reduction
Phase 5 (Documentation & Polish):
- README: architecture overview, feature table, examples guide
- Public API docs on all exported types and methods
- SAFETY comments on all 46+ unsafe blocks across 5 files
- Panics docs, init() technical docs
- Cargo doc: 0 warnings from our crate (escaped brackets, qualified links)
- Fix static_mut_refs warning in mass_storage example
- Suppress QTD_TOKEN_CPAGE_SHIFT dead_code with annotation

Phase 6 (Code Organization & Unsafe Reduction):
- Split host.rs (2308 lines) into 10 focused submodules
- UnsafeCell-based DMA pools (eliminate const-to-mut UB)
- QtdSlot RAII guard (auto-free qTDs, eliminate error-path cleanup)
- Schedule link+flush helpers (deduplicate async schedule code)
- VCell interior mutability fix (UnsafeCell, eliminate raw volatile bypass)
- Reduce unsafe impl Send: 3 eliminated via u32 base addresses / pool indices
- Separate qTD allocation bitmap from hardware token field
- New hub enumeration example (rtic_usb_hub.rs)
- Zero warnings from our crate in both build and cargo doc
